refactor(*): refactor SDK and add tests * add unit tests * update SDK structure * clean analyzer fields * add unix processor * add utilities
