+++ .golangci.yml
@@ -1,3 +1,3 @@
 run:
   skip-dirs:
-    - sdk/triggers
+    - analyzers/triggers
