+++ analyzers/testdata/triggers/staticcheck/staticcheck.go
